oooooooooooooo

+++ migrations/20220604012623-create_conversao_table.js
@@ -10,11 +10,12 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       valor_original: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT
       },
       valor_convertido: {
-        type: Sequelize.INTEGER
-      }
+        type: Sequelize.FLOAT
+      },
+     
     });
   },
